fix: updateOption api dispose chartInstance
