+++ packages/vtable/src/layout/pivot-layout.ts
@@ -1560,4 +1560,11 @@ export class PivotLayoutMap implements LayoutMapAPI {
     return false;
   }
   //#endregion
+  dispose() {
+    const activeChartInstance = (this._table as PivotChart)._getActiveChartInstance();
+    activeChartInstance?.release();
+    this._indicatorObjects.forEach(indicatorObject => {
+      indicatorObject.chartInstance?.release();
+    });
+  }
 }
