+++ packages/vtable/src/layout/simple-header-layout.ts
@@ -739,4 +739,11 @@ export class SimpleHeaderLayoutMap implements LayoutMapAPI {
   getChartDataId(col: number, row: number): any {
     return getChartDataId(col, row, this);
   }
+  dispose() {
+    const activeChartInstance = this._table._getActiveChartInstance();
+    activeChartInstance?.release();
+    this.columnObjects.forEach(indicatorObject => {
+      indicatorObject.chartInstance?.release();
+    });
+  }
 }
